Typing and documentation changes
