+++ pint/facets/context/definitions.py
@@ -1,5 +1,5 @@
 """
-    pint.facets.systems.definitions
+    pint.facets.context.definitions
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     :copyright: 2022 by Pint Authors, see AUTHORS for more details.
