+++ pint/facets/nonmultiplicative/objects.py
@@ -1,3 +1,11 @@
+"""
+    pint.facets.nonmultiplicative.objects
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    :copyright: 2022 by Pint Authors, see AUTHORS for more details.
+    :license: BSD, see LICENSE for more details.
+"""
+
 from __future__ import annotations
 
 from typing import List
