+++ pint/facets/numpy/quantity.py
@@ -1,3 +1,11 @@
+"""
+    pint.facets.numpy.quantity
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    :copyright: 2022 by Pint Authors, see AUTHORS for more details.
+    :license: BSD, see LICENSE for more details.
+"""
+
 from __future__ import annotations
 
 import functools
